requirements.txt
psycopg2-binary
numpy
matplotlib
pandas
seaborn
scikit-learn